categoryName feauture hashing

+++ streamlit_app.py
@@ -64,7 +64,7 @@ n_features = len(df.categoryName.unique())
 categories = df.categoryName.astype(str)  
 categories = [[category] for category in categories]
 hasher = FeatureHasher(n_features=n_features, input_type="string")
-X_category = hasher.transform(categories).toarray().astype("float32")
+X_category = hasher.transform(categories).toarray().astype("float16")
 hashed_df = pd.DataFrame(X_category, columns=[f"hash_{i}" for i in range(n_features)])
 
 # Concatenate dataframes
